addinventorytoorder integration tests
