+++ tests/GardeningExpress.DespatchCloudClient.Tests.Integration/CreateOrderAsyncTests.cs
@@ -1,5 +1,6 @@
 ﻿using GardeningExpress.DespatchCloudClient.DTO.Request;
 using GardeningExpress.DespatchCloudClient.Model.GoGroopie;
+using GardeningExpress.DespatchCloudClient.Tests.Integration.Utils;
 using Newtonsoft.Json;
 using NUnit.Framework;
 using System.Collections.Generic;
@@ -20,49 +21,7 @@ namespace GardeningExpress.DespatchCloudClient.Tests.Integration
         [SetUp]
         public void Setup()
         {
-            newOrder = new OrderCreateRequest()
-            {
-                StatusId = 1,
-                ShippingMethodRequested = "Test",
-                PaymentMethod = "CC",
-                PaymentRef = "CC-123",
-                PaymentCurrency = "GBP",
-                TotalPaid = "100",
-                TotalDiscount = "0",
-                TotalTax = "0",
-                TotalWeight = "100",
-                Email = "demo@despatchcloud.com",
-                PhoneOne = "+905559998877",
-                VatNumber = "V1",
-                EoriNumber = "E1",
-                TaxId = "T1",
-                ShippingName = "Test Customer 2",
-                ShippingAddressLineOne = "SAL 1",
-                ShippingAddressLineTwo = "SAL 2",
-                ShippingAddressCity = "Izmir",
-                ShippingAddressCounty = "Ege",
-                ShippingAddressCountry = "Turkey",
-                ShippingAddressPostcode = "35090",
-                ShippingAddressIso = "TR",
-                InvoiceName = "Invoice Name 1",
-                InvoiceAddressLineOne = "IAL 1",
-                InvoiceAddressLineTwo = "IAL 2",
-                InvoiceAddressCity = "Bursa",
-                InvoiceAddressCounty = "Marmara",
-                InvoiceAddressCountry = "Turkey",
-                InvoiceAddressPostcode = "16300",
-                InvoiceAddressIso = "TR",
-                CustomerComments = "Test 1 Comment",
-                Notes = new List<string> {
-                    "Test 1 Note"
-                },
-                CustomFields = new Dictionary<string, string>() {
-                    { "test1-1","Test Data" }
-                },
-                FulfilmentClientId = null,
-                ShippingPaid = "2.00",
-                ManualChannelId = 1
-            };
+            newOrder = TestUtils.GetCreateOrderRequest();
         }
                 
         [Test]
@@ -80,8 +39,8 @@ namespace GardeningExpress.DespatchCloudClient.Tests.Integration
         }
 
         [Theory]
-        [TestCase("PaymentMethod", null)]
-        public async Task CreateOrderAsync_ShouldReturnIsSuccessAsFalse_WhenRequiredFieldMissing(string field, object setValue = null)
+        [TestCase("PaymentMethod", "payment_method", null)]
+        public async Task CreateOrderAsync_ShouldReturnIsSuccessAsFalseAndAnErrorMessage_WhenRequiredFieldMissing(string field, string jsonField, object setValue = null)
         {
             // ARRANGE
             TestUtils.SetPropertyValue(newOrder, field, setValue);
@@ -91,6 +50,7 @@ namespace GardeningExpress.DespatchCloudClient.Tests.Integration
 
             // ASSERT
             Assert.IsFalse(result.IsSuccess);
+            Assert.AreEqual(result.Error, $"Error executing statement: Column '{jsonField}' cannot be null");
         }
 
         [Test]
@@ -105,6 +65,8 @@ namespace GardeningExpress.DespatchCloudClient.Tests.Integration
 
             // ASSERT
             Assert.IsTrue(result.IsSuccess);
+            Assert.NotNull(result.Data);
+            Assert.NotNull(result.Data.Id); // Verify trackable id set
         }
     }
 }
